Refactor and enhance core features and database schema
- Added new endpoints across multiple controllers, including `CheckStatus`, `GetQuestionsByTest`, and `SubmitAnswer`.
- Updated DTOs and validators to support new properties and nullable fields.
- Replaced `Username` with `Email` for authentication and user identification.
- Refactored repository methods for flexible query customization.
- Enhanced `TestResultService` with transaction support and new methods.
- Introduced new DTOs, services, and AutoMapper profiles for modularity.
- Updated database schema with migrations to support new features:
  - Added new columns (`ShortDescription`, `DetailedExplanation`, `Recommendations`) to `ScoreRange`.
  - Adjusted cascade delete behaviors for better data integrity.
- Improved seeding process with predefined tests and cleanup logic.
- Enhanced validation, error handling, and transaction management.
- Removed unused code, dependencies, and mapping profiles for maintainability.
- Updated test cases to align with domain model and DTO changes.

+++ ExamiNation.API/Controllers/Test/ScoreRangeController.cs
@@ -56,7 +56,7 @@ namespace ExamiNation.API.Controllers.Test
 
         [Authorize(Roles = RoleGroups.AdminOrDevOrCreator)]
         [HttpGet("classification")]
-        public async Task<IActionResult> GetClassificationAsync([FromQuery] Guid testId, [FromQuery] int score)
+        public async Task<IActionResult> GetClassificationAsync([FromQuery] Guid testId, [FromQuery] decimal score)
         {
             if (testId == Guid.Empty)
             {
